Added Rosenbrock methods
