+++ lib/OrdinaryDiffEqRosenbrock/Project.toml
@@ -1,0 +1,20 @@
+name = "OrdinaryDiffEqRosenbrock"
+uuid = "849cf443-90a5-45fe-9245-010b279924ce"
+authors = ["ParamThakkar123 <paramthakkar864@gmail.com>"]
+version = "1.0.0"
+
+[deps]
+OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+MuladdMacro = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
+
+[extras]
+DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+julia = "1.10"
+
+[targets]
+test = ["DiffEqDevTools", "Random", "SafeTestsets", "Test"]
