+++ lib/OrdinaryDiffEqRosenbrock/src/interp_func.jl
@@ -1,0 +1,19 @@
+function DiffEqBase.interp_summary(::Type{cacheType},
+        dense::Bool) where {
+        cacheType <:
+        Union{Rosenbrock23ConstantCache,
+        Rosenbrock32ConstantCache,
+        Rosenbrock23Cache,
+        Rosenbrock32Cache}}
+    dense ? "specialized 2nd order \"free\" stiffness-aware interpolation" :
+    "1st order linear"
+end
+
+function DiffEqBase.interp_summary(::Type{cacheType},
+        dense::Bool) where {
+        cacheType <:
+        Union{Rosenbrock5ConstantCache,
+        Rosenbrock5Cache}}
+    dense ? "specialized 4rd order \"free\" stiffness-aware interpolation" :
+    "1st order linear"
+end
