+++ lib/OrdinaryDiffEqRosenbrock/test/runtests.jl
@@ -1,0 +1,3 @@
+using SafeTestsets
+
+@time @safetestset "Rosenbrock Tests" include("ode_rosenbrock_tests.jl")
