+++ src/alg_utils.jl
@@ -488,42 +488,6 @@ alg_order(alg::Hairer4) = 4
 alg_order(alg::Hairer42) = 4
 alg_order(alg::PFRK87) = 8
 
-alg_order(alg::ROS2) = 2
-alg_order(alg::ROS2PR) = 2
-alg_order(alg::ROS2S) = 2
-alg_order(alg::ROS3) = 3
-alg_order(alg::ROS3PR) = 3
-alg_order(alg::Scholz4_7) = 3
-alg_order(alg::Rosenbrock23) = 2
-alg_order(alg::Rodas23W) = 3
-alg_order(alg::Rosenbrock32) = 3
-alg_order(alg::ROS3P) = 3
-alg_order(alg::Rodas3) = 3
-alg_order(alg::Rodas3P) = 3
-alg_order(alg::ROS34PW1a) = 3
-alg_order(alg::ROS34PW1b) = 3
-alg_order(alg::ROS34PW2) = 3
-alg_order(alg::ROS34PW3) = 4
-alg_order(alg::ROS34PRw) = 3
-alg_order(alg::ROS3PRL) = 3
-alg_order(alg::ROS3PRL2) = 3
-alg_order(alg::ROK4a) = 4
-alg_order(alg::RosShamp4) = 4
-alg_order(alg::Veldd4) = 4
-alg_order(alg::Velds4) = 4
-alg_order(alg::GRK4T) = 4
-alg_order(alg::GRK4A) = 4
-alg_order(alg::Ros4LStab) = 4
-alg_order(alg::RosenbrockW6S4OS) = 4
-alg_order(alg::Rodas4) = 4
-alg_order(alg::Rodas42) = 4
-alg_order(alg::Rodas4P) = 4
-alg_order(alg::Rodas4P2) = 4
-alg_order(alg::Rodas5) = 5
-alg_order(alg::Rodas5P) = 5
-alg_order(alg::Rodas5Pr) = 5
-alg_order(alg::Rodas5Pe) = 5
-
 alg_order(alg::AB3) = 3
 alg_order(alg::AB4) = 4
 alg_order(alg::AB5) = 5
@@ -573,9 +537,6 @@ alg_maximum_order(alg::CompositeAlgorithm) = maximum(alg_order(x) for x in alg.a
 
 alg_adaptive_order(alg::ExplicitRK) = alg.tableau.adaptiveorder
 alg_adaptive_order(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = alg_order(alg) - 1
-
-alg_adaptive_order(alg::Rosenbrock23) = 3
-alg_adaptive_order(alg::Rosenbrock32) = 2
 
 alg_adaptive_order(alg::RadauIIA3) = 1
 alg_adaptive_order(alg::RadauIIA5) = 3
@@ -738,17 +699,6 @@ isstandard(alg::OrdinaryDiffEqNewtonAdaptiveAlgorithm) = alg.controller === :Sta
 isstandard(alg::VCABM) = true
 
 isWmethod(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = false
-isWmethod(alg::Rosenbrock23) = true
-isWmethod(alg::Rosenbrock32) = true
-isWmethod(alg::Rodas23W) = true
-isWmethod(alg::ROS2S) = true
-isWmethod(alg::ROS34PW1a) = true
-isWmethod(alg::ROS34PW1b) = true
-isWmethod(alg::ROS34PW2) = true
-isWmethod(alg::ROS34PW3) = true
-isWmethod(alg::ROS34PRw) = true
-isWmethod(alg::ROK4a) = true
-isWmethod(alg::RosenbrockW6S4OS) = true
 
 isesdirk(alg::TRBDF2) = true
 function isesdirk(alg::Union{KenCarp3, KenCarp4, KenCarp5, KenCarp58,
@@ -761,7 +711,6 @@ isesdirk(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = false
 
 is_mass_matrix_alg(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = false
 is_mass_matrix_alg(alg::CompositeAlgorithm) = all(is_mass_matrix_alg, alg.algs)
-is_mass_matrix_alg(alg::RosenbrockAlgorithm) = true
 is_mass_matrix_alg(alg::NewtonAlgorithm) = !isesdirk(alg)
 # hack for the default alg
 function is_mass_matrix_alg(alg::CompositeAlgorithm{
