+++ src/interp_func.jl
@@ -48,30 +48,11 @@ end
 function DiffEqBase.interp_summary(::Type{cacheType},
         dense::Bool) where {
         cacheType <:
-        Union{Rosenbrock23ConstantCache,
-        Rosenbrock32ConstantCache,
-        Rosenbrock23Cache,
-        Rosenbrock32Cache}}
-    dense ? "specialized 2nd order \"free\" stiffness-aware interpolation" :
-    "1st order linear"
-end
-function DiffEqBase.interp_summary(::Type{cacheType},
-        dense::Bool) where {
-        cacheType <:
         Union{Rodas4ConstantCache, Rodas23WConstantCache, Rodas3PConstantCache,
         Rodas4Cache, Rodas23WCache, Rodas3PCache}}
     dense ? "specialized 3rd order \"free\" stiffness-aware interpolation" :
     "1st order linear"
 end
-function DiffEqBase.interp_summary(::Type{cacheType},
-        dense::Bool) where {
-        cacheType <:
-        Union{Rosenbrock5ConstantCache,
-        Rosenbrock5Cache}}
-    dense ? "specialized 4rd order \"free\" stiffness-aware interpolation" :
-    "1st order linear"
-end
-
 function DiffEqBase.interp_summary(::Type{cacheType},
         dense::Bool) where {
         cacheType <: Union{OwrenZen3Cache,
