+++ test/runtests.jl
@@ -201,11 +201,14 @@ end
     if !is_APPVEYOR && GROUP == "AlgConvergence_III"
         @time @safetestset "Linear Methods Tests" include("algconvergence/linear_method_tests.jl")
         @time @safetestset "Split Methods Tests" include("algconvergence/split_methods_tests.jl")
-        @time @safetestset "Rosenbrock Tests" include("algconvergence/ode_rosenbrock_tests.jl")
         @time @safetestset "FIRK Tests" include("algconvergence/ode_firk_tests.jl")
         @time @safetestset "Linear-Nonlinear Methods Tests" include("algconvergence/linear_nonlinear_convergence_tests.jl")
         @time @safetestset "Linear-Nonlinear Krylov Methods Tests" include("algconvergence/linear_nonlinear_krylov_tests.jl")
         @time @safetestset "Quadruple precision Runge-Kutta Tests" include("algconvergence/ode_quadruple_precision_tests.jl")
+    end
+
+    if !is_APPVEYOR && GROUP == "Rosenbrock"
+        @time @safetestset "Rosenbrock Tests" include("../lib/OrdinaryDiffEqRosenbrock/test/ode_rosenbrock_tests.jl")
     end
 
     if !is_APPVEYOR && GROUP == "Symplectic"
